--- src/frontend/src/views/cart/EmptyCart.vue
@@ -1,22 +1,0 @@
-<template>
-  <form action="#" method="post" class="layout-form">
-    <main class="content cart">
-      <div class="container">
-        <div class="cart__title">
-          <h1 class="title title--big">Корзина</h1>
-        </div>
-        <div class="sheet cart__empty">
-          <p>В корзине нет ни одного товара</p>
-        </div>
-      </div>
-    </main>
-  </form>
-</template>
-
-<script>
-export default {
-  name: "EmptyCart",
-};
-</script>
-
-<style scoped></style>
